Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    float a,b;
+    int a,b;
     char opr;
-    scanf("%f %f %c", &a, &b, &opr);
+    scanf("%d %d %c", &a, &b, &opr);
     if(opr=='+'){
         printf("%.0f",a+b);
     }
